+++ commands/build.go
@@ -32,17 +32,14 @@ const postBuildScriptName = "postbuild"
 const preBuildScriptName = "prebuild"
 
 func Init_Build_Command(parentCmd *cobra.Command, app *types.AppContext) {
-	var noScript bool
-	var noPostScript bool
-	var noPreScript bool
-
 	var buildCmd = &cobra.Command{
 		Use:     "build",
 		Aliases: []string{"b"},
 		Short:   "Runs build command",
 		Long:    `Runs the 'build' script or the official 'go build .'.`,
 		Run: func(cmd *cobra.Command, args []string) {
-			if !noPreScript {
+			if !app.NoPreScript {
+				// prebuild defined?
 				_, ok := app.GpmFile.Scripts[preBuildScriptName]
 				if ok {
 					app.RunScript(preBuildScriptName)
@@ -50,7 +47,7 @@ func Init_Build_Command(parentCmd *cobra.Command, app *types.AppContext) {
 			}
 
 			_, ok := app.GpmFile.Scripts[buildScriptName]
-			if !noScript && ok {
+			if !app.NoScript && ok {
 				app.RunScript(buildScriptName, args...)
 			} else {
 				cmdArgs := []string{"go", "build", "."}
@@ -59,7 +56,8 @@ func Init_Build_Command(parentCmd *cobra.Command, app *types.AppContext) {
 				app.RunShellCommandByArgs(cmdArgs[0], cmdArgs[1:]...)
 			}
 
-			if !noPostScript {
+			if !app.NoPostScript {
+				// postbuild defined?
 				_, ok := app.GpmFile.Scripts[postBuildScriptName]
 				if ok {
 					app.RunScript(postBuildScriptName)
@@ -67,10 +65,6 @@ func Init_Build_Command(parentCmd *cobra.Command, app *types.AppContext) {
 			}
 		},
 	}
-
-	buildCmd.Flags().BoolVarP(&noScript, "no-script", "n", false, "do not handle '"+buildScriptName+"' script")
-	buildCmd.Flags().BoolVarP(&noPostScript, "no-post-script", "", false, "do not handle '"+postBuildScriptName+"' script")
-	buildCmd.Flags().BoolVarP(&noPreScript, "no-pre-script", "", false, "do not handle '"+preBuildScriptName+"' script")
 
 	parentCmd.AddCommand(
 		buildCmd,